[수정] 팀원 평가 로직 수정

+++ src/main/java/com/studycollaboproject/scope/service/TeamService.java
@@ -3,6 +3,7 @@ package com.studycollaboproject.scope.service;
 import com.studycollaboproject.scope.dto.MemberListResponseDto;
 import com.studycollaboproject.scope.exception.ErrorCode;
 import com.studycollaboproject.scope.exception.BadRequestException;
+import com.studycollaboproject.scope.exception.ForbiddenException;
 import com.studycollaboproject.scope.model.*;
 import com.studycollaboproject.scope.repository.ApplicantRepository;
 import com.studycollaboproject.scope.repository.TeamRepository;
@@ -50,21 +51,20 @@ public class TeamService {
                 .stream().map(MemberListResponseDto::new).collect(Collectors.toCollection(ArrayList::new));
     }
 
-    public void memberResignation(User member, Post post) {
-        if (post.getProjectStatus().equals(ProjectStatus.PROJECT_STATUS_RECRUITMENT)) {
-            teamRepository.deleteByUserAndPost(member, post);
-        } else throw new BadRequestException(ErrorCode.NO_RECRUITMENT_ERROR);
-    }
-
-    public void memberSecession(User user, Post post) {
-
-        if (post.getProjectStatus().equals(ProjectStatus.PROJECT_STATUS_RECRUITMENT)) {
+    public void memberDelete(User user, Post post) {
+        if (post.getUser().getId().equals(user.getId())) {
+            throw new BadRequestException(ErrorCode.NOT_AVAILABLE_ACCESS);
+        }
+        else if (post.getProjectStatus().equals(ProjectStatus.PROJECT_STATUS_RECRUITMENT)) {
+            Team team = teamRepository.findByUserAndPost(user, post).orElseThrow(
+                    () -> new ForbiddenException(ErrorCode.NO_TEAM_ERROR)
+            );
+            team.deleteTeam();
             teamRepository.deleteByUserAndPost(user, post);
         } else throw new BadRequestException(ErrorCode.NO_RECRUITMENT_ERROR);
-
     }
 
-    public boolean isMemeber(Post post, User user) {
+    public boolean isMember(Post post, User user) {
         return teamRepository.findByUserAndPost(user,post).isPresent();
 
     }
